Fix partition value type casting
The partition field is mapped as a String and is cast as a BigDecimal to
get the integer value. The problem is that it's not possible to cast a
String to a BigDecimal with usual type casting, this type of casting
ends up throwing a ClassCastException.

BigDecimal is supposed to represent precision signed decimal numbers,
but the partition is never going to be a decimal or unsigned number, so
the casting was changed to parse directory to an integer.